requirements.txt
accelerate==1.12.0
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.13.2
filelock==3.20.3
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ipykernel==7.1.0
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain-core==1.2.9
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.6.9
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==2.17.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
parso==0.8.5
peft==0.18.1
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.3
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.1.0
triton==3.6.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
wcwidth==0.5.3
xxhash==3.6.0
zstandard==0.25.0
